pyproject.toml
[tool.poetry]
authors     = [ "Ruzzy77 <hiyaq77@gmail.com>" ]
description = ""
name        = "color-grabbers"
readme      = "README.md"
version     = "0.1.0"

    [tool.poetry.dependencies]
    fastapi                 = "^0.115.6"
    pandas                  = "^2.2.3"
    pillow                  = "^11.0.0"
    pyshorteners            = "^1.0.1"
    python                  = "^3.12"
    qrcode                  = "^8.0"
    streamlit               = "^1.41.1"
    streamlit-authenticator = "^0.4.1"
    supabase                = "^2.10.0"
    tabulate                = "^0.9.0"
    uvicorn                 = "^0.34.0"
    st-theme = "^1.2.3"


[build-system]
build-backend = "poetry.core.masonry.api"
requires      = [ "poetry-core" ]
